Change HTKMLFReader to Work With I-Vectors in Write Mode
I-vector feature files are one frame long, but should be replicated to every
frame of the speech segment. This change will, for every feature stream
after the first, expand single-frame segments to match the length of the
first feature stream. This change only affects the writer mode of
the HTKMLFReader.